Include the contents of connectivity.h into tria.cc.
